requirements.txt
Flask==3.0.3
Jinja2==3.1.4
python-dotenv==1.0.1
google-api-python-client==2.136.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
gunicorn==22.0.0
requests==2.32.3
